requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
coreap==2.3.3
djangorestframewor==3.14.0
djangorestframework-simplejwt==5.3.0